split code
split validation func, controllers to proper files
