+++ app.js
@@ -19,6 +19,7 @@ app.use((req, res) => {
 })
 
 app.use((err, req, res, next) => {
+  console.log('err', err)
      res.status(err.status).json(err.message);
 })
