+++ models/contacts.json
@@ -28,5 +28,11 @@
     "name": "Zebra",
     "email": "scelerisque@magnis.org",
     "phone": "(186) 568-3720"
+  },
+  {
+    "id": "wIEEWbxeW2mjC04svXBT6",
+    "name": "Kennedy Lane the Second",
+    "email": "mattis.Cras@nonenimMauris.net.ua",
+    "phone": "(999) 888-0000"
   }
 ]
